+++ internal/scout/service_test.go
@@ -56,30 +56,6 @@ func TestCalculateDiagnostics(t *testing.T) {
 					},
 					Severity: types.CreateDiagnosticSeverityPointer(protocol.DiagnosticSeverityWarning),
 				},
-				{
-					Message: "Tag recommendations available",
-					Code:    &protocol.IntegerOrString{Value: "recommended_tag"},
-					Source:  types.CreateStringPointer("scout-testing-source"),
-					Range: protocol.Range{
-						Start: protocol.Position{Line: 0, Character: 0},
-						End:   protocol.Position{Line: 0, Character: 18},
-					},
-					Severity: types.CreateDiagnosticSeverityPointer(protocol.DiagnosticSeverityInformation),
-					Data: []types.NamedEdit{
-						{
-							Title: "Update image to preferred tag (3.21.3)",
-							Edit:  "FROM alpine:3.21.3",
-						},
-						{
-							Title: "Update image OS minor version (3.20.6)",
-							Edit:  "FROM alpine:3.20.6",
-						},
-						{
-							Title: "Update image OS minor version (3.18.12)",
-							Edit:  "FROM alpine:3.18.12",
-						},
-					},
-				},
 			},
 		},
 		{
@@ -98,30 +74,6 @@ func TestCalculateDiagnostics(t *testing.T) {
 						End:   protocol.Position{Line: 0, Character: 44},
 					},
 					Severity: types.CreateDiagnosticSeverityPointer(protocol.DiagnosticSeverityWarning),
-				},
-				{
-					Message: "Tag recommendations available",
-					Code:    &protocol.IntegerOrString{Value: "recommended_tag"},
-					Source:  types.CreateStringPointer("scout-testing-source"),
-					Range: protocol.Range{
-						Start: protocol.Position{Line: 0, Character: 0},
-						End:   protocol.Position{Line: 0, Character: 44},
-					},
-					Severity: types.CreateDiagnosticSeverityPointer(protocol.DiagnosticSeverityInformation),
-					Data: []types.NamedEdit{
-						{
-							Title: "Update image to preferred tag (3.21.3)",
-							Edit:  "FROM --platform=$BUILDPLATFORM alpine:3.21.3",
-						},
-						{
-							Title: "Update image OS minor version (3.20.6)",
-							Edit:  "FROM --platform=$BUILDPLATFORM alpine:3.20.6",
-						},
-						{
-							Title: "Update image OS minor version (3.18.12)",
-							Edit:  "FROM --platform=$BUILDPLATFORM alpine:3.18.12",
-						},
-					},
 				},
 			},
 		},
@@ -210,14 +162,12 @@ func TestCalculateDiagnostics_IgnoresSpecifics(t *testing.T) {
 			content: "FROM alpine:3.16.1",
 			codes: []string{
 				"critical_high_vulnerabilities",
-				"recommended_tag",
 			},
 		},
 		{
 			name:    "ubuntu:24.04",
 			content: "FROM ubuntu:24.04",
 			codes: []string{
-				"recommended_tag",
 				"vulnerabilities",
 			},
 		},
@@ -326,7 +276,7 @@ func TestGetHovers(t *testing.T) {
 			result: &protocol.Hover{
 				Contents: protocol.MarkupContent{
 					Kind:  protocol.MarkupKindMarkdown,
-					Value: "Current image vulnerabilities:   1C   3H   9M   0L \r\n\r\nRecommended tags:\n\n<table>\n<tr><td><code>3.21.3</code></td><td align=\"right\">  0C</td><td align=\"right\">  0H</td><td align=\"right\">  0M</td><td align=\"right\">  0L</td><td align=\"right\"></td></tr>\n<tr><td><code>3.20.6</code></td><td align=\"right\">  0C</td><td align=\"right\">  0H</td><td align=\"right\">  0M</td><td align=\"right\">  0L</td><td align=\"right\"></td></tr>\n<tr><td><code>3.18.12</code></td><td align=\"right\">  0C</td><td align=\"right\">  0H</td><td align=\"right\">  0M</td><td align=\"right\">  0L</td><td align=\"right\"></td></tr>\n</table>\n",
+					Value: "Current image vulnerabilities:   1C   3H   9M   0L ",
 				},
 			},
 		},