+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to the Docker Language Server will be documented in this fil
 
 ### Fixed
 
+- correct initialize request handling for clients that do not support dynamic registrations ([#229](https://github.com/docker/docker-language-server/issues/229))
 - Dockerfile
   - textDocument/hover
     - hide vulnerability hovers if the top level setting is disabled ([#226](https://github.com/docker/docker-language-server/issues/226))