+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the Docker Language Server will be documented in this file.
 
-## [Unreleased]
+## [0.9.0] - 2025-05-26
 
 ## Added
 
@@ -328,7 +328,8 @@ All notable changes to the Docker Language Server will be documented in this fil
   - textDocument/semanticTokens/full
     - provide syntax highlighting for Bake files
 
-[Unreleased]: https://github.com/docker/docker-language-server/compare/v0.8.0...main
+[Unreleased]: https://github.com/docker/docker-language-server/compare/v0.9.0...main
+[0.9.0]: https://github.com/docker/docker-language-server/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/docker/docker-language-server/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/docker/docker-language-server/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/docker/docker-language-server/compare/v0.5.0...v0.6.0